Basic Route Sover, Config file, and Graph Library
This is the baseline, it can solve for a single path or all possible paths. Allows for the user to either provide a config file at command line, or assemble the graph manually. I was unable to add the directional parameter to this iteration. Will be on the next commit, as well as picking the best route out of all routes.