Changes to run the pi comms throughout the route
Several changes in this commit:
Moved all files to the /pi/ directory for ease of import
Created a new route_solver.py file that is a condensed version of the original command line driven one. (Still uses yaml #TODO)
updated the pi_I2c file to take in the map, and end node from the server comms and start the driving/continue along the route through to finish.